Split `main` module up into `gps` and `map`
This should give a good example of how one can use RequireJS, and helps
improve code readability by grouping related parts together better.
